refactor(cache): seperate marshaller and compressot
