+++ cache/compressor/compressor.go
@@ -1,0 +1,8 @@
+package compressor
+
+import (
+	"github.com/alexfalkowski/go-service/compressor"
+)
+
+// Compressor for cache.
+type Compressor compressor.Compressor
