+++ cache/compressor/snappy.go
@@ -1,0 +1,11 @@
+package compressor
+
+import (
+	"github.com/alexfalkowski/go-service/compressor"
+)
+
+// NewSnappy for cache.
+// nolint:ireturn
+func NewSnappy() Compressor {
+	return compressor.NewSnappy()
+}
