+++ cache/marshaller/marshaller.go
@@ -1,0 +1,8 @@
+package marshaller
+
+import (
+	"github.com/alexfalkowski/go-service/marshaller"
+)
+
+// Marshaller for cache.
+type Marshaller marshaller.Marshaller
