+++ cache/marshaller/proto.go
@@ -1,0 +1,11 @@
+package marshaller
+
+import (
+	"github.com/alexfalkowski/go-service/marshaller"
+)
+
+// NewProto for cache.
+// nolint:ireturn
+func NewProto() Marshaller {
+	return marshaller.NewProto()
+}
