+++ cache/module.go
@@ -1,6 +1,8 @@
 package cache
 
 import (
+	"github.com/alexfalkowski/go-service/cache/compressor"
+	"github.com/alexfalkowski/go-service/cache/marshaller"
 	"github.com/alexfalkowski/go-service/cache/redis"
 	"github.com/alexfalkowski/go-service/cache/ristretto"
 	"go.uber.org/fx"
@@ -14,4 +16,12 @@ var (
 	// RistrettoModule for fx.
 	// nolint:gochecknoglobals
 	RistrettoModule = fx.Options(fx.Provide(ristretto.NewCache))
+
+	// SnappyCompressorModule for fx.
+	// nolint:gochecknoglobals
+	SnappyCompressorModule = fx.Options(fx.Provide(compressor.NewSnappy))
+
+	// ProtoMarshallerModule for fx.
+	// nolint:gochecknoglobals
+	ProtoMarshallerModule = fx.Options(fx.Provide(marshaller.NewProto))
 )
