+++ cmd/cmd_test.go
@@ -12,7 +12,6 @@ import (
 	"github.com/alexfalkowski/go-service/cache/redis"
 	"github.com/alexfalkowski/go-service/cache/ristretto"
 	"github.com/alexfalkowski/go-service/cmd"
-	"github.com/alexfalkowski/go-service/compressor"
 	"github.com/alexfalkowski/go-service/config"
 	"github.com/alexfalkowski/go-service/database/sql"
 	"github.com/alexfalkowski/go-service/database/sql/pg"
@@ -20,7 +19,6 @@ import (
 	hgrpc "github.com/alexfalkowski/go-service/health/transport/grpc"
 	hhttp "github.com/alexfalkowski/go-service/health/transport/http"
 	"github.com/alexfalkowski/go-service/logger"
-	"github.com/alexfalkowski/go-service/marshaller"
 	"github.com/alexfalkowski/go-service/security"
 	"github.com/alexfalkowski/go-service/security/auth0"
 	"github.com/alexfalkowski/go-service/trace"
@@ -44,7 +42,7 @@ func TestShutdown(t *testing.T) {
 				config.Module, logger.ZapModule, health.GRPCModule, health.HTTPModule, health.ServerModule,
 				cache.RedisModule, cache.RistrettoModule, security.Auth0Module, sql.PostgreSQLModule,
 				trace.JaegerOpenTracingModule, transport.HTTPServerModule, transport.GRPCServerModule,
-				marshaller.ProtoModule, compressor.SnappyModule,
+				cache.ProtoMarshallerModule, cache.SnappyCompressorModule,
 				fx.Provide(registrations), fx.Provide(healthObserver), fx.Provide(livenessObserver),
 				fx.Provide(readinessObserver), fx.Provide(grpcObserver), fx.Invoke(shutdown), fx.Invoke(configs),
 			}
@@ -71,7 +69,7 @@ func TestInvalidHTTP(t *testing.T) {
 				fx.NopLogger,
 				config.Module, logger.ZapModule, transport.HTTPServerModule, transport.GRPCServerModule,
 				health.GRPCModule, health.HTTPModule, health.ServerModule, trace.JaegerOpenTracingModule,
-				marshaller.ProtoModule, compressor.SnappyModule,
+				cache.ProtoMarshallerModule, cache.SnappyCompressorModule,
 				fx.Provide(registrations), fx.Provide(healthObserver), fx.Provide(livenessObserver),
 				fx.Provide(readinessObserver), fx.Provide(grpcObserver),
 			}
@@ -101,7 +99,7 @@ func TestInvalidGRPC(t *testing.T) {
 				fx.NopLogger,
 				config.Module, logger.ZapModule, transport.HTTPServerModule, transport.GRPCServerModule,
 				health.GRPCModule, health.HTTPModule, health.ServerModule, trace.DataDogOpenTracingModule,
-				marshaller.ProtoModule, compressor.SnappyModule,
+				cache.ProtoMarshallerModule, cache.SnappyCompressorModule,
 				fx.Provide(registrations), fx.Provide(healthObserver), fx.Provide(livenessObserver),
 				fx.Provide(readinessObserver), fx.Provide(grpcObserver),
 			}
@@ -146,7 +144,7 @@ func TestInvalidClient(t *testing.T) {
 				fx.NopLogger,
 				config.Module, logger.ZapModule, transport.HTTPServerModule, transport.GRPCServerModule,
 				health.GRPCModule, health.HTTPModule, health.ServerModule, trace.DataDogOpenTracingModule,
-				marshaller.ProtoModule, compressor.SnappyModule,
+				cache.ProtoMarshallerModule, cache.SnappyCompressorModule,
 				fx.Provide(registrations), fx.Provide(healthObserver), fx.Provide(livenessObserver),
 				fx.Provide(readinessObserver), fx.Provide(grpcObserver),
 			}
