--- compressor/module.go
@@ -1,9 +1,0 @@
-package compressor
-
-import (
-	"go.uber.org/fx"
-)
-
-// SnappyModule for fx.
-// nolint:gochecknoglobals
-var SnappyModule = fx.Options(fx.Provide(NewSnappy))
